fix unauthorised requests
currently if a request is unauthorised to kong then gokong would return
a `nil` error and a `nil` resource.  The `v1.8.0` release of `gokong`
fixes this and now a non `nil` error is returned in this case.

fixed a small bug in the if statement when creating an api resource to
stop a nil pointer exception from being possible.

fixes #18
fixes #34