Use GetOkExists for reading pointer types (#94)
* failing test for setting a service's "retries" to zero

* switch from GetOk to GetOkExists for pointer types

Terraform's GetOk treats zero values the same as being not set. For
readIntPtrFromResource and readStringPtrFromResource, this prevents us
from explicitly setting a value to 0 (or the empty string) because the
code believes nothing has been set (rather than an explicit zero). This
commit fixes that by using GetOkExists, which has more predictable
semantics