Implement `kong_consumer_key_auth` resource
By adapting the previously copied `kong_consumer_basic_auth` resource.

Note that the TTL property is deliberately omitted from the resource,
because the value is not stable - when retrieved from Kong the value
returned is the remaining TTL based on when it was created, so it's not
very well suited to Terraforming.

Possibly it could be a computed value, but will be left out for now.