tests: `kong_target` should be recreated if removed
Added this test as I noticed that this resource also doesn't contain the check `!kong.IsNotFoundErr(err)`. However, this resource works slightly differently in that the Kong library returns a blank list with no error if no targets are found. And so, the test passed. Regardless, I figured I may as well leave the test it in place just in case this behaviour changes in future.